linux-tinker-v-rzfive: userspace workaround for local memory
Within the LM (Local Memory) address space, the virtual address
will be treated as a physical address and will not trigger page
fault. While executing a user program, it is possible to encounter
the segmentation fault.

Unfortunately, this functionality cannot be disabled via any
custom CSR. It is crucial to restrict userspace from touching
virtual addresses ranging from 0x30000 to 0x4ffff.

The native binutils is also been modified to adjust the text base
of user programs. This modification enables the gcc installed on
the board to perform static linking with a safe text base address.

Signed-off-by: Yu Chien Peter Lin <peterlin@andestech.com>